+++ test/generators/active_storage_generators/azure_active_storage_install_generator_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "generators/boring/active_storage/aws/install/install_generator"
+require "generators/boring/active_storage/azure/install/install_generator"
 
-class AwsActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
-  tests Boring::ActiveStorage::Aws::InstallGenerator
+class AzureActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
+  tests Boring::ActiveStorage::Azure::InstallGenerator
   setup :build_app
   teardown :teardown_app
 
@@ -14,16 +14,16 @@ class AwsActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
     app_path
   end
 
-  def test_should_configure_aws_for_active_storage
+  def test_should_configure_azure_for_active_storage
     Dir.chdir(app_path) do
       quietly { run_generator }
 
       assert_file "Gemfile" do |content|
-        assert_match(/aws-sdk-s3/, content)
+        assert_match(/azure-storage-blob/, content)
       end
 
       assert_file "config/environments/production.rb" do |content|
-        assert_match("config.active_storage.service = :amazon", content)
+        assert_match("config.active_storage.service = :microsoft", content)
         assert_no_match("config.active_storage.service = :local", content)
       end
     end